Remove Rust MCP server
I want to continue with a setup that enables more developer efficiency.
This means Typescript and Solid.js for the UI and Python for MCP server
and CLI. Both will provide a better environment for quick experiments
and fast iterations.

Once the fucntionality is more settled, rewriting the core functionality
in Rust is an obvious and sustainable solution. But not now!

Signed-off-by: Jan Ehrhardt <59441+jehrhardt@users.noreply.github.com>